Centralize config in `config.rs`
- Move all WriterConfig, ReaderConfig, and CompactorConfig to `config.rs`.
- Add a smidge of documentation to `flush_ms`.
- Repalce the compactor poll interval constant with a default constant to make
  it look like the other defaults. This required a `.clone()` in the bench test
  because the Tokio executor clones itself (and takes ownership of the options
  lifetime).
- Update `compactor_execute_bench.rs` to use `DEFAULT_COMPACTOR_OPTIONS`.
